Try to avoid nerfstudio import when training
